+++ src/components/Overlay/Overlay.js
@@ -1,33 +1,9 @@
-import React from "react";
-import ReactDOM from "react-dom";
 import styled from "styled-components";
 import { prop } from "styled-tools";
 import as from "../../enhancers/as";
 import Hidden from "../Hidden";
 
-class Component extends React.Component {
-  state = {};
-
-  componentDidMount() {
-    const wrapper = document.createElement("div");
-    document.body.appendChild(wrapper);
-    this.setState({ wrapper });
-  }
-
-  componentWillUnmount() {
-    document.body.removeChild(this.state.wrapper);
-  }
-
-  render() {
-    const { wrapper } = this.state;
-    if (wrapper) {
-      return ReactDOM.createPortal(<Hidden {...this.props} />, wrapper);
-    }
-    return null;
-  }
-}
-
-const Overlay = styled(Component)`
+const Overlay = styled(Hidden)`
   position: fixed;
   background-color: white;
   left: 50%;
